#  VMS-7855 Reset to defaults button
# Test to cach crash from VMS-8178
# VMS-7855 bug fixes
# VMS-7855 minor stability fixes
# VMS-7855 more fixes
# VMS-7855 real impl
# VMS-8082 QnPhysicalCameraResource -> nx::mediaserver::resource::Camera
# VMS-8082 bug fix
# VMS-8082 get rid of get/set in nx::mediaserver::resource::Camera inheritors
# VMS-8082 minor bug fixes
# VMS-8082 more fixes
# VMS-8082 move out advanced parameters providers
# VMS-8082 nx::mediaserver::resource::Camera::AdvancedParametersProvider's
#VMS-7855 advanced params page don't display initial value correctly
VMS-7258 Remove extra debug output from LayoutThumbnailLoader
VMS-7358 Another fix for background in ExportSettingsDialog
VMS-7855: Introduce streams advanced parameters for all cameras.
VMS-7855: bug fixing
VMS-7928: Added DW-Edge motion properties
VMS-7928: Enabling dw-edge software motion detection for cmake
VMS-7928: Fixing Alexey's commit
VMS-8050: Introduce method for fetching resource media stream capabilities.
VMS-8053: DW H265 fix
VMS-8053: DW driver fix
VMS-8053: Update all openStream methods so it takes all parameters from QnLiveStreamParameters
VMS-8053: add UT
VMS-8053: add database migration
VMS-8053: deprecate code is removed from expert tab
VMS-8053: disableDualStreaming UI setting fix
VMS-8053: enable bitrate control for all vendors on schedule tab
VMS-8053: fix documentation, rename signal
VMS-8053: fix fps
VMS-8053: race condition is fixed
VMS-8055: do not write empty parameters to settings
VMS-8055: make separate tabs for settings and web page inside 'Advanced' tab
VMS-8070 Added QnSecurityCamResource::audioEnabledChanged signal
VMS-8070 Added missing include
VMS-8070 Aligner for wearable archive length widget
VMS-8070 Asserting in upload worker added
VMS-8070 Better naming
VMS-8070 Better single camera settings dialog
VMS-8070 Better single camera widget.
VMS-8070 Better stop uploads message box
VMS-8070 Better wearable archive sync task
VMS-8070 Better wearable camera addition dialog
VMS-8070 Build =(
VMS-8070 Drop -Async suffix in rest connection methods
VMS-8070 Dropped excessive this->
VMS-8070 Evil managers ftw
VMS-8070 Explicitly clear workers in upload manager's dtor
VMS-8070 Field initialization in declaration for archive length widget
VMS-8070 FileUpload in a separate .h file.
VMS-8070 Filename instead of path in wearable progress widget
VMS-8070 Fix crash in single camera settings dialog
VMS-8070 Fix destruction order for upload/wearable managers
VMS-8070 Hide event rules & dignose buttons for wearable cameras
VMS-8070 Hide uploadFolder button for now
VMS-8070 Less boilerplate in upload worker
VMS-8070 Linux build
VMS-8070 More safety in wearable manager
VMS-8070 No live stream overlay
VMS-8070 Progress in archive sync task
VMS-8070 Proper exit handling
VMS-8070 Proper include for FileUpload.
VMS-8070 Properly muting wearable camera when audio is disabled
VMS-8070 Race fixed
VMS-8070 Remote archive worker pool now also cleans up queue on cancel
VMS-8070 Remote archive worker pool supports adding tasks directly
VMS-8070 Renamed FileUpload -> UploadState
VMS-8070 Saner upload manager API
VMS-8070 Separate progress & buttons wearable widgets.
VMS-8070 Server request storage
VMS-8070 Simpler wearable progress widget code
VMS-8070 Some cleanup in wearable worker
VMS-8070 Style
VMS-8070 TODO added
VMS-8070 Validator for new wearable camera dialog
VMS-8070 Waaay simpler remove archive synchronizer
VMS-8070 Wearable archive synchronizer
VMS-8070 Wearable manager can now manage with cameras being deleted right under its nose
VMS-8070 Wearable sync threads global setting
VMS-8070 Wearable upload widget now polls for state updates
VMS-8070 WearableManager api to get / cancel all uploads
VMS-8070 WearableWorker::isWorking
VMS-8095: New DW multi-directional cameras are added with stithced view
VMS-8115: do not reset UPNP cache if auto discovery is turned off
VMS-8115: do not reset UPNP cache if discovery is not enabled
VMS-8115: move Acti searcher to new UPNP class. Get rid of /devicedesc.xml request
VMS-8126: Get rid of SecondaryStreamQuality
VMS-8168 hide wearable controls for normal cameras.
VMS-8170 Reopen stream on advanced video paramiters change
VMS-8171 Fixed: cloud_connectivity_ut/ConnectorExecutor* tests did not cleaned up after themselves
VMS-8173: Crash detected in server: bb827dfc7d9e0571be756cbd80462a7e
VMS-8193: partially reverted commit ec620105bca7
VMS-8195: Fix removing previously added USB storage from server's
VMS-8235: Advanced parameter string should be the same each time
VMS-CR-8612 VMS-8171: Compilation fix
VMS-CR-8612 VMS-8171: Fixed crash in MediatorConnector.* unit tests
VMS-CR-8612 VMS-8171: MediatorConnector: Added delay before reconnecting to mediator
VMS-CR-8692 VMS-8171: Fixed duplicate 'connection closed' event reporting in AsyncClientWithHttpTunneling class. Added more tests.

<==== Mercurial data ====>
Branch: vms

+++ mediaserver_core/src/rtsp/rtsp_connection.cpp
@@ -873,9 +873,10 @@ int QnRtspConnectionProcessor::composeDescribe()
         auto mediaStreamIter = std::find_if(
             supportedMediaStreams.streams.cbegin(),
             supportedMediaStreams.streams.cend(),
-            [mediaStreamIndex]( const CameraMediaStreamInfo& streamInfo ) {
-                return streamInfo.encoderIndex == mediaStreamIndex;
-            } );
+            [mediaStreamIndex]( const CameraMediaStreamInfo& streamInfo )
+            {
+                return (int) streamInfo.encoderIndex == mediaStreamIndex;
+            });
         const std::map<QString, QString>& streamParams =
             mediaStreamIter != supportedMediaStreams.streams.cend()
             ? mediaStreamIter->customStreamParams
